Fix formatting issues in HTML templates

+++ CW2/app/views.py
@@ -1,7 +1,7 @@
-from app import app,db,login_manager,api
-from flask import render_template, flash, redirect,request,jsonify,url_for
+from app import app, db, login_manager, api
+from flask import render_template, flash, redirect, request, jsonify, url_for
 from datetime import datetime
-from .forms import LoginForm, UserForm, PostForm,UpdateAccountForm,CommentForm,ChangePasswordForm,deleteAccountForm,DeletePostForm, UpdatePostForm,DeletePostForm,DeleteCommentForm
+from .forms import LoginForm, UserForm, PostForm, UpdateAccountForm, CommentForm, ChangePasswordForm, deleteAccountForm, DeletePostForm, UpdatePostForm, DeletePostForm, DeleteCommentForm
 from .models import User, Post, Tag, post_tag, Comment
 from flask_login import login_user, logout_user, login_required, current_user
 from werkzeug.utils import secure_filename
@@ -9,7 +9,6 @@ from werkzeug.security import generate_password_hash, check_password_hash
 import uuid as uuid
 from sqlalchemy import func
 import re
-
 
 
 @app.route('/')
@@ -20,7 +19,7 @@ def home():
 @app.route('/register', methods=['GET', 'POST'])
 def register():
     register_form = UserForm()
-    if(register_form.validate_on_submit()):
+    if (register_form.validate_on_submit()):
         fname = register_form.fname.data
         lname = register_form.lname.data
         username = register_form.username.data
@@ -28,14 +27,15 @@ def register():
         password = register_form.password.data
         # Check if the user exists in the database
         user = User.query.filter_by(email=email).first()
-        if(user is not None):
+        if (user is not None):
             print("User exists")
             # User exists, redirect to the login page
             return redirect('/login')
         else:
             print("User does not exist")
             # User does not exist, create a new user
-            user = User(name=fname +' '+ lname, email=email, username = username, password=password)
+            user = User(name=fname + ' ' + lname, email=email,
+                        username=username, password=password)
             # Add the new user to the database
             db.session.add(user)
             db.session.commit()
@@ -48,7 +48,7 @@ def register():
             # Redirect to the login page
             return redirect('/login')
 
-    return render_template("register.html",register_form = register_form)
+    return render_template("register.html", register_form=register_form)
 
 
 @app.route('/register/validate-username', methods=['POST'])
@@ -60,12 +60,12 @@ def validate_username():
         return jsonify({'success': True})
     else:
         return jsonify({'success': False, 'error': 'Username already exists'})
-    
-    
+
 
 @login_manager.user_loader
 def load_user(id):
     return User.query.get(int(id))
+
 
 @app.route('/login', methods=['GET', 'POST'])
 def login():
@@ -77,16 +77,16 @@ def login():
         password = login_form.password.data
         # Using the username and password query the database to see if the user exists
         user = User.query.filter_by(username=username).first()
-        if(user is not None and user.verify_password(password)):
+        if (user is not None and user.verify_password(password)):
             # Log in the user
             login_user(user)
-            if(user.username == 'admin'):
+            if (user.username == 'admin'):
                 return redirect('/admin')
             else:
                 return redirect('/explore')
         else:
             return redirect('/login')
-    return render_template("login.html",login_form = login_form)
+    return render_template("login.html", login_form=login_form)
 
 
 @app.route('/dashboard', methods=['GET', 'POST'])
@@ -96,23 +96,23 @@ def userDash():
     register_form = UpdateAccountForm()
     password_form = ChangePasswordForm()
     delete_form = deleteAccountForm()
-    if 'update' in request.form :
+    if 'update' in request.form:
         # Check if username is changed and not unique
         if (register_form.username.data != user.username and
-            User.query.filter(User.username == register_form.username.data).first()):
+                User.query.filter(User.username == register_form.username.data).first()):
             flash("Username already in use. Please choose another one.", "error")
             return redirect(url_for('userDash'))
 
         # Check if email is changed and not unique
         if (register_form.email.data != user.email and
-            User.query.filter(User.email == register_form.email.data).first()):
+                User.query.filter(User.email == register_form.email.data).first()):
             flash("Email already in use. Please choose another one.", "error")
             return redirect(url_for('userDash'))
 
         # Check if the name is changed
         if register_form.name.data:
             user.name = register_form.name.data
-        
+
         user.name = register_form.name.data
         user.username = register_form.username.data
         user.email = register_form.email.data
@@ -120,23 +120,24 @@ def userDash():
         db.session.commit()
         flash('Your account has been updated!', 'success')
         return redirect(url_for('userDash'))
-    
+
     elif 'delete' in request.form:
         print("Delete")
-        if  'delete' in request.form and current_user.username != 'admin':
+        if 'delete' in request.form and current_user.username != 'admin':
             if check_password_hash(current_user.passwordHashed, delete_form.confirm_password.data):
-                #Delete the comments of the user
-                comments = Comment.query.filter_by(author_id=current_user.id).all()
+                # Delete the comments of the user
+                comments = Comment.query.filter_by(
+                    author_id=current_user.id).all()
                 for comment in comments:
                     db.session.delete(comment)
                     db.session.commit()
-                #Delete the posts of the user
+                # Delete the posts of the user
                 posts = Post.query.filter_by(author_id=current_user.id).all()
                 for post in posts:
                     db.session.delete(post)
                     db.session.commit()
-                
-                # Delete the user   
+
+                # Delete the user
                 db.session.delete(user)
                 db.session.commit()
                 flash('Your account has been deleted!', 'success')
@@ -163,9 +164,9 @@ def userDash():
         flash('Your password has been updated!', 'success')
         return redirect(url_for('userDash'))
 
-
     print(request.form)
-    return render_template("user_dash.html", user=user, register_form=register_form,delete_form=delete_form,password_form=password_form)
+    return render_template("user_dash.html", user=user, register_form=register_form, delete_form=delete_form, password_form=password_form)
+
 
 @app.route('/my_page', methods=['GET', 'POST'])
 @login_required
@@ -174,32 +175,32 @@ def my_page():
     posts = Post.query.filter_by(author_id=current_user.id).all()
     latest_posts = []  # Initialize the variable before the loop
     for _ in posts:
-        latest_posts = Post.query.filter_by(author_id=current_user.id).order_by(Post.publish_date.desc()).limit(5).all()
+        latest_posts = Post.query.filter_by(author_id=current_user.id).order_by(
+            Post.publish_date.desc()).limit(5).all()
     # Return all of the posts of the user except the post with the highest number of comments and the latest posts
     top_three_posts = Post.query \
-    .outerjoin(Comment, Post.id == Comment.post_id) \
-    .with_entities(Post, func.count(Comment.id).label('comment_count')) \
-    .group_by(Post.id) \
-    .order_by(func.count(Comment.id).desc()) \
-    .limit(3) \
-    .all()
-    
-     #Display the number of posts and comments of the user 
+        .outerjoin(Comment, Post.id == Comment.post_id) \
+        .with_entities(Post, func.count(Comment.id).label('comment_count')) \
+        .group_by(Post.id) \
+        .order_by(func.count(Comment.id).desc()) \
+        .limit(3) \
+        .all()
+
+    # Display the number of posts and comments of the user
     posts = Post.query.filter_by(author_id=current_user.id).all()
     comments = Comment.query.filter_by(author_id=current_user.id).all()
     post_count = len(posts)
     comment_count = len(comments)
-    
-    #Display the number of tags used by the user
+
+    # Display the number of tags used by the user
     tags = Tag.query.all()
     tag_count = 0
     for tag in tags:
         for post in posts:
             if tag in post.tags:
                 tag_count += 1
-                
-    return render_template("my_page.html",user=user, posts = posts, latest_posts = latest_posts, remaining_posts = top_three_posts,tag_count=tag_count,post_count=post_count,comment_count=comment_count)
 
+    return render_template("my_page.html", user=user, posts=posts, latest_posts=latest_posts, remaining_posts=top_three_posts, tag_count=tag_count, post_count=post_count, comment_count=comment_count)
 
 
 @app.route('/logout', methods=['GET'])
@@ -213,14 +214,14 @@ def logout():
 @app.route('/addPost', methods=['GET', 'POST'])
 def add_post():
     post_form = PostForm()
-    if(request.method == 'POST'):
+    if (request.method == 'POST'):
         # Get the data from the form
         title = request.form['title']
         # Extract post content and tags from the form
         content = request.form['caption']
         # Check if the title is not duplicate from the same user
         post = Post.query.filter_by(title=title).first()
-        if(post is not None):
+        if (post is not None):
             print("Post exists, change the title")
             # Post exists, redirect to the login page
             return redirect('/addPost')
@@ -235,42 +236,41 @@ def add_post():
                 db.session.add(tag)
             # No need to commit yet, it will happen after adding to post
             tag_instances.append(tag)
-    
+
         # Now, create a new Post instance
-        post = Post(title = title,caption=content, author_id=current_user.id)
+        post = Post(title=title, caption=content, author_id=current_user.id)
         db.session.add(post)
-    
+
         # Add each tag to the post
         for tag in tag_instances:
             post.tags.append(tag)
-    
+
         # Commit once to insert everything
         db.session.commit()
         flash('Your post has been created! , Check it out in the explore page ')
-    print ("Hello")
-    return render_template("add_post.html",post_form = post_form)
-
+    print("Hello")
+    return render_template("add_post.html", post_form=post_form)
 
 
 @app.route('/explore', methods=['GET', 'POST'])
 def explore():
     # Store the users in a dictionary for easy lookup
     userCache = {}
-    # get the page number from the request 
+    # get the page number from the request
     page = request.args.get('page', 1, type=int)
     # Paginate the posts
-    posts = Post.query.order_by(Post.publish_date.desc()).paginate(page=page, per_page=4)
+    posts = Post.query.order_by(
+        Post.publish_date.desc()).paginate(page=page, per_page=4)
     for post in posts.items:
         if post.author_id in userCache:
             authorName = userCache[post.author_id]
         else:
-            #Fetch the author from the database
+            # Fetch the author from the database
             author = User.query.filter_by(id=post.author_id).first()
             # Add the author to the cache
             userCache[post.author_id] = author.name
 
-    return render_template("explore.html", posts = posts.items , userCache = userCache, paginationPost = posts)
-
+    return render_template("explore.html", posts=posts.items, userCache=userCache, paginationPost=posts)
 
 
 @app.route('/post/<int:post_id>', methods=['GET', 'POST'])
@@ -283,7 +283,6 @@ def post(post_id):
     delete_post_form = DeletePostForm()
     delete_comment_form = DeleteCommentForm()
     updatePost = UpdatePostForm()
-    updateComment = UpdateCommentForm()
     userCache = {user.id: user.name for user in User.query.all()}
 
     if request.method == 'POST':
@@ -321,10 +320,12 @@ def post(post_id):
             )
             db.session.add(new_comment)
             db.session.commit()
-            return redirect(url_for('post', post_id=post_id))  # Redirect after form submission
+            # Redirect after form submission
+            return redirect(url_for('post', post_id=post_id))
     current_user_id = current_user.id
     # Regular GET request
-    comments_with_authors = db.session.query(Comment, User.name).join(User, User.id == Comment.author_id).filter(Comment.post_id == post_id).all()
+    comments_with_authors = db.session.query(Comment, User.name).join(
+        User, User.id == Comment.author_id).filter(Comment.post_id == post_id).all()
     return render_template(
         "post_page.html",
         post=post,
@@ -336,24 +337,23 @@ def post(post_id):
         comments_with_authors=comments_with_authors,
         delete_post_form=delete_post_form,
         delete_comment_form=delete_comment_form,
-        updatePost=updatePost,
-        updateComment=updateComment
+        updatePost=updatePost
     )
-    
 
-@app.route('/delete_comment/<int:comment_id>', methods=['POST'])  
+
+@app.route('/delete_comment/<int:comment_id>', methods=['POST'])
 @login_required
 def delete_comment(comment_id):
     comment = Comment.query.get_or_404(comment_id)
     if (comment.author_id != current_user.id) or (current_user.username == 'admin'):
         flash('You cannot delete this comment.', 'danger')
-        return redirect(url_for('index'))
+        return redirect(url_for('explore'))
     else:
         db.session.delete(comment)
-        db.session.commit()  
+        db.session.commit()
         flash('Your comment has been deleted!', 'success')
     return redirect(url_for('post', post_id=comment.post_id))
-    
+
 
 @app.route('/edit_post/<int:post_id>', methods=['POST'])
 @login_required
@@ -361,15 +361,17 @@ def edit_post(post_id):
     post = Post.query.get_or_404(post_id)
     if post.author_id != current_user.id:
         flash('You cannot edit this post.', 'danger')
-        return redirect(url_for('index'))  # Redirect to the homepage or other appropriate page
+        # Redirect to the homepage or other appropriate page
+        return redirect(url_for('index'))
 
-    post_form = PostForm(obj=post)  # Pre-populate the form with the existing post
-    
+    # Pre-populate the form with the existing post
+    post_form = PostForm(obj=post)
+
     if request.method == "POST":
         # Update the post's title and content
         post.title = post_form.title.data
         post.caption = post_form.caption.data
-        
+
         # Clear existing tags and add the new ones from the form content
         post.tags.clear()
         tags = set(re.findall(r'#([A-Za-z0-9_]+)', post.caption))
@@ -383,9 +385,9 @@ def edit_post(post_id):
         # Commit the changes to the database
         db.session.commit()
         flash('Your post has been updated!', 'success')
-        return redirect(url_for('post', post_id=post.id))  # Redirect to the view post page
+        # Redirect to the view post page
+        return redirect(url_for('post', post_id=post.id))
     return redirect(url_for('post', post_id=post_id))
-
 
 
 @app.route('/delete_post/<int:post_id>', methods=['GET', 'POST'])
@@ -393,7 +395,7 @@ def edit_post(post_id):
 def delete_post(post_id):
     delete_post_form = DeletePostForm()
     post = Post.query.get_or_404(post_id)
-    if (request.method == 'POST' and post.author_id == current_user.id) or (request.method == 'POST' and current_user.username == 'admin'):
+    if (request.method == 'POST' and post.author_id == current_user.id) or (request.method == 'POST' and current_user.username == 'Admin'):
         # Delete all of the comments of the post
         comments = Comment.query.filter_by(post_id=post_id).all()
         for comment in comments:
@@ -407,17 +409,25 @@ def delete_post(post_id):
     flash('You can only delete your own posts!', 'danger')
     return redirect(url_for('post', post_id=post_id))
 
+
 @app.route('/admin', methods=['GET', 'POST'])
 @login_required
 def admin():
+    if current_user.username != "Admin":  # Assuming "admin" is the admin's username
+        flash('You are not authorized to access this page.', 'danger')
+        return redirect(url_for('explore'))  # Make sure to use return here!
+
     # The admin views the users and views the number of posts and comments of each user
     users = User.query.filter(User.id != 1).all()
     # Load the delete user form
     delete_user_form = deleteAccountForm()
-     # Create dictionaries to hold the post and comment counts for each user
-    user_posts_counts = {user.id: Post.query.filter_by(author_id=user.id).count() for user in users}
-    user_comments_counts = {user.id: Comment.query.filter_by(author_id=user.id).count() for user in users}
-    return render_template("admin_dash.html",users=users,user_posts_counts=user_posts_counts,user_comments_counts=user_comments_counts,delete_user_form=delete_user_form )
+    # Create dictionaries to hold the post and comment counts for each user
+    user_posts_counts = {user.id: Post.query.filter_by(
+        author_id=user.id).count() for user in users}
+    user_comments_counts = {user.id: Comment.query.filter_by(
+        author_id=user.id).count() for user in users}
+    return render_template("admin_dash.html", users=users, user_posts_counts=user_posts_counts, user_comments_counts=user_comments_counts, delete_user_form=delete_user_form)
+
 
 @app.route('/admin/delete_user/<int:user_id>', methods=['GET', 'POST'])
 @login_required
@@ -444,4 +454,3 @@ def deleteUsers(user_id):
     db.session.commit()
     flash('User deleted successfully!', 'success')
     return redirect(url_for('admin'))
-
